+++ Assets/Scripts/Character/Enemy/CharacterEnemyNormal.cs
@@ -1,4 +1,3 @@
-using System.Numerics;
 using UnityEngine;
 using System.Collections;
 
@@ -59,7 +58,7 @@ public class CharacterEnemyNormal : CharacterEnemy
 
     public override void Dead()
     {
-        BigDecimal reward = UpgradeManager.Instance.GetUpgrade("Reward").currentValue / (6.0f * 8.0f);
+        float reward = UpgradeManager.Instance.GetUpgrade("Reward").currentValue / (6.0f * 8.0f);
         for(int i = 0; i < 8; ++i)
         {
             ObjectPool<ItemGold>.Spawn("@ItemGold", position + new Vector3(0.0f, 1.0f)).Init(reward);
